# Rework to remove implicit flushes