--- a/examples/client.rs
+++ b/examples/client.rs
@@ -14,9 +14,9 @@ fn main() {
         println!("* {}", header);
     }
 
-    socket.write_message(Message::Text("Hello WebSocket".into())).unwrap();
+    socket.send(Message::Text("Hello WebSocket".into())).unwrap();
     loop {
-        let msg = socket.read_message().expect("Error reading message");
+        let msg = socket.read().expect("Error reading message");
         println!("Received: {}", msg);
     }
     // socket.close(None);