--- a/tests/receive_after_init_close.rs
+++ b/tests/receive_after_init_close.rs
@@ -29,12 +29,12 @@ fn test_receive_after_init_close() {
     let client_thread = spawn(move || {
         let (mut client, _) = connect(Url::parse("ws://localhost:3013/socket").unwrap()).unwrap();
 
-        client.write_message(Message::Text("Hello WebSocket".into())).unwrap();
+        client.send(Message::Text("Hello WebSocket".into())).unwrap();
 
-        let message = client.read_message().unwrap(); // receive close from server
+        let message = client.read().unwrap(); // receive close from server
         assert!(message.is_close());
 
-        let err = client.read_message().unwrap_err(); // now we should get ConnectionClosed
+        let err = client.read().unwrap_err(); // now we should get ConnectionClosed
         match err {
             Error::ConnectionClosed => {}
             _ => panic!("unexpected error: {:?}", err),
@@ -47,12 +47,12 @@ fn test_receive_after_init_close() {
     client_handler.close(None).unwrap(); // send close to client
 
     // This read should succeed even though we already initiated a close
-    let message = client_handler.read_message().unwrap();
+    let message = client_handler.read().unwrap();
     assert_eq!(message.into_data(), b"Hello WebSocket");
 
-    assert!(client_handler.read_message().unwrap().is_close()); // receive acknowledgement
+    assert!(client_handler.read().unwrap().is_close()); // receive acknowledgement
 
-    let err = client_handler.read_message().unwrap_err(); // now we should get ConnectionClosed
+    let err = client_handler.read().unwrap_err(); // now we should get ConnectionClosed
     match err {
         Error::ConnectionClosed => {}
         _ => panic!("unexpected error: {:?}", err),