My context: I've been working on a websocket load generator service that will write lots of small json message events to it's ws clients. I was interested in having a high peak throughput from this service to allow me to test downstream load performance. It uses axum->tokio-tungstenite->tungstenite. Doing localhost testing I got **~1.2M msg/s** peak throughput. This PR improves that to **~2M msg/s**.

Because of this improvement I thought this PR may be worthwhile.

### Summary: Reduce flushing and allow better control of when flushes happen
I noticed there was a lot of flushing happening. Currently flushing is done something like:
* Before writing each message.
* After writing each message.
* Before reading each message.

This seems excessive particularly when trying to send batches of lots of small messages. This PR minimises flushes so that a flush will only occur if `write_pending` is called, or if a "system message" (pong/close) should be dispatched. So users can call `write_message` many times without flushing (which is more like how I expected it to behave), then call `write_pending` which will flush.

### Impl notes
My first approach was to remove the flush before each frame write currently in `WebSocketContext::write_pending`
```rust
// First, make sure we have no pending frame sending.
self.frame.write_pending(stream)?;
```
And remove the flush after each `FrameCodec::write_frame`. Then wire up `write_message` to **not** call flush and `write_pending` to essentially _be_ flush.

#### Flush before write & `send_queue`
I noticed that after these changes the `WebSocketContext::send_queue` no longer really seemed to do much. The logic as is actually **requires** flushing before each write to work otherwise a message is simply added to the `send_queue` then taken off it and put in the `FrameCodec::out_buffer`.

So I removed the `send_queue` since it no longer does anything. It didn't seem to do too much previously, as it could only build up there if flushing before each write returned an error. If the write itself returned an error the send_queue would be empty and the message will simply be stored in the `FrameCodec::out_buffer`. I'm not sure what the benefit of having both a send_queue and an out_buffer is/was?

This PR I replace the `send_queue` by using the existing `out_buffer`. Writes are buffered there without flushing before each. We can buffer many writes here, as was done before, but without any send queue.

I replaced the `max_send_queue` config with `max_write_buffer_size` which can bound the previously unbound `out_buffer`. If a write would exceed this bound the message is instead returned as an error, in a very similar way previous logic would handle the `send_queue` bound. The one difference here is the messages is returned as a `Message::Frame` since it has already been serialized at this point. This seems worse than before where the message was returned as it was given. However, logic to re-send should still work. I'm not sure this difference is worth the flush call to prevent. It might be possible to put in a conversion back to the old message somehow but I'm not sure it's worth it.

#### Flush after each write
Removing this seems straightforward. We now require the caller calls `write_pending` after any writes to guarantee it will be flushed. This requirement seems similar to most write interfaces.

An exception is "system messages" like pong responses. If these have been scheduled we'll still automatically flush.

#### Flush on each read
Currently `WebSocketContext::read_message` will flush **writes** on each call.
```rust
// Since we may get ping or close, we need to reply to the messages even during read.
// Thus we call write_pending() but ignore its blocking.
self.write_pending(stream).no_block()?;
```

This means that if my service is polling for incoming messages I can still end up flushing writes a lot more than I'd expect. This PR changes this to only write if we actually have a pong/close "system message" to send. Since system messages won't scale with load this minimises extra flush impact while still writing-and-flushing these messages asap.

### micro-bench
I added a micro-bench `benches/write.rs` to reproduce flush cost. This simulates a `Write` impl that takes 100ns to flush. The scenario is 100k `write_message`s then one `write_pending`. E.g. how I would approach writing a batch of small messages with the current API.

The improvement is clear since this PR will call only one flush per batch.
```
write 100k small texts then flush
        time:   [6.1961 ms 6.2280 ms 6.2360 ms]
        change: [-66.921% -66.617% -66.309%] (p = 0.07 > 0.05)
```

### Issues
* The micro-bench is quite artificial. Perhaps it could be made more realistic without too much bloat?
* `Error::WriteBufferFull` will return the message as a `Message::Frame`, whereas `SendQueueFull` would return the message as given.
* This is a **breaking** change.
* It may be worth considering some API changes to reflect the changes. E.g. rename `write_pending` to `flush`?