@agalakhov 2023-05-26
Thank you very much for the improvement! Technically this looks fine. But there are two small issues.

First, does automatic Pong reply still works correctly? It used to rely on automatic flush while reading. Also, does it work correctly if there is already some message in outgoing buffer? I'm quite sure I introduced the outgoing queue mostly for handling pings correctly.

Second, I'd rather rename `write_message` to something else in order to break old code. There are probably pieces of code here and there on the web that still rely on implicit flushing, so let's make them non-compileable. Renaming `write_pending` to `flush` is a good idea as well.