@alexheretic 2023-05-26
> First, does automatic Pong reply still works correctly? It used to rely on automatic flush while reading. Also, does it work correctly if there is already some message in outgoing buffer? I'm quite sure I introduced the outgoing queue mostly for handling pings correctly.

Auto pong replies should still auto flush.

Previously `read_message`->`read_message_frame` would set `WebSocketContext::pong` to Some. The next call to `write_message`, `write_pending` or `read_message` would write+flush since they all flushed writes. I don't think the send_queue was used for this, and indeed deduping auto-pongs is desirable.

In this PR `read_message`->`read_message_frame` will set `WebSocketContext::additional_send` to Some in a similar way (generalised to allow setting close frames here too). As before the next call to `write_message`, `write_pending` or `read_message` will write+flush the pong. The difference is calls to `write_message` & `read_message` **won't** flush in the general case where `WebSocketContext::additional_send` is None.

So we flush eagerly whenever we have a auto generated pong or close to send otherwise we only flush on user call.

Related: If the user sends a **manual pong** there is no implicit flushing, since if the user writes they can also flush as normal. My thinking was implicit/auto flushing only really made sense for auto generated messages outside of the user's direct control.

> Second, I'd rather rename write_message to something else in order to break old code. There are probably pieces of code here and there on the web that still rely on implicit flushing, so let's make them non-compileable. Renaming write_pending to flush is a good idea as well.

That's a good idea. Perhaps we have methods `write` & `flush` so to align with `Write` trait? There could also be another method that simply does both write+flush, a bit like [`SinkExt::send` vs `feed`](https://docs.rs/futures/latest/futures/sink/trait.SinkExt.html#method.send). This could be called `send`. I'm not 100% sure how useful that is downstream, just an idea.

We could also keep a deprecated `write_message` & `write_pending` that both work similarly as before (write+flush) to ease migration. Then presumably remove the deprecated code on the _next_ breaking change.