@agalakhov 2023-05-26
Yes, `write` and `flush` sounds good. Then probably `read` as well. And one has to change `tokio_tungstenite` accordingly.

How does Pong work now if there is something not yet flushed and a Ping is incoming? This is the only thing that has to be verified. Probably it is covered by Autobahn test suite but I'd like to double check this before merging.