@alexheretic 2023-05-27
> Yes, write and flush sounds good. Then probably read as well. And one has to change tokio_tungstenite accordingly.

Ok I'll make these modifications when I get a moment.

> How does Pong work now if there is something not yet flushed and a Ping is incoming?

Once a ping is read a pong is scheduled to be written and flushed on the next call to read or write. On that call, pong is appended to the `out_buffer` all of the `out_buffer` will be written and flushed. So if stuff already resides in that buffer, e.g. because of a previous write error, the pong frame is just appended on top and we try to write & flush all of it.