@alexheretic 2023-05-27
I have renamed the methods to better reflect the changes. We now use `read`, `send`, `write` & `flush`. `send` is write+flush and seems nice for simple usage and for migrating previous usage.

I've added entries to the changelog.

To aid migration I have deprecated `WebSocket::read_message`, `write_message` & `write_pending` and deprecated `WebSocketConfig::max_send_queue`. This should prevent many users from seeing "hard" breaks and the warnings should be easy to action. I _didn't_ do the same for `WebSocketContext` or `FrameSocket` as they seem lower level, so less of an issue.