@alexheretic 2023-05-28
Great to see merged, thanks for reviewing!

I've prepared https://github.com/snapview/tokio-tungstenite/pull/284 for the merged `0.20` API. Do you have an idea of when the new release could be published?